add methods for searching related posts
The all-posts component can now find relevant and related posts based on
the tags defined in the frontmatter of each mdx posts. The amount of
common tags is a measure of how relevant a post is to the currently
viewed post. When calculated relationship between post clashes, the
posts are arranged in a list, and the order in which they appear is
determined by their last modified.