Caleb (#1)
* initial commit - ammendment

* Import our global styles

* Gatsby configuration and plugins

* SCSS: sass script variables

* sass script for our cards component

* card component

* container component

* footer component

* sass script for footer

* front-tile component used on the home page

* removed

* A component that enumerates all our posts so far

* header component

* our global styles sass script

* Our home/default layout

* layout component for our posts

* seo component for setting meta tags and seo opt.

* sidebar components

* sass script for our sidebar component

* Image query example component

* blog-wide images

* 404 NotFound component

* first blog post

* Create gatsby-node.js

This file contains middlewares for the gatsby compile time. Activities that should be run at build time, before build starts or after build ends, are put here

* our index/home page

* Delete favicon.ico

* added path clipping and new post plus favicon

* correct typo

* add methods for searching related posts

The all-posts component can now find relevant and related posts based on
the tags defined in the frontmatter of each mdx posts. The amount of
common tags is a measure of how relevant a post is to the currently
viewed post. When calculated relationship between post clashes, the
posts are arranged in a list, and the order in which they appear is
determined by their last modified.

* site images

* change description and add offline support plugin

* npm package data

* display front tile only on desktop screen sizes

* rewrite authors href to remove unecessary "@" char

* modify site wide margins for h2 - h6 elements

* make footer sticky to bottom of page

* remove irrelevant links

* include helper function and modify TOC margins

* add support for json data loading

By installing necessary gatsby plugins for json and settting up the
gatsby config to use the plugin.

* adjust toc depth count to reflect well ordered pad

* layout for pages

* default layout: remove FrontTile component

* new page

* button component

* modified

* add new nav item

add featured post to nav links and add css styling for active nav

* posts

* images

* index page