reload disqus on theme switch
When the theme switches, `DiscussionEmbedded` Disqus component doesn't
adapt to the new theme until a reload occurs.
This is fixed by adding a hook that generates a static key for the
current theme.
This key is passed as a `key` prop to the disqus component, so whenever
the key changes the component rerenders, causing it to reload and update.